Make transposition table thread-safe
The transposition table can now be shared across threads using Arc.
This is the first step towards implementing lazy SMP. The perfomance
impact is almost negligible compared to the old implementation of the
transposition table. Of course, we should get a nice speedup when
we start actually running on multiple threads.