Add Lazy SMP
Rewrite most of the solver module to be able to run with multiple
threads. Number of threads can be set in the CLI.

NOTE: Currently running with more threads slows down the search. I
suspect that this is because the move ordering does not take advantage
of the transposition table. This essentially means that each thread is
just going through the search space in the exact same order.
Furthermore, since the transposition table is shared, if one thread is
lagging behind, it will quickly catch up to the other threads, causing
them to finish at nearly the exact same time. Hopefully, storing the best
column in the transposition table and using it for move ordering will
make the Lazy SMP actually work...